Expand the PIC and weak symbol macros on the assembly files
The GNU assembler will not accept multiple instructions on a single line
so all multiline macros have to be expanded.  It was also unclear that the
"common" assembly files are i386-only assembly, so trying to execute them
on amd64 will result in weird errors because the macros are set wrong in
that case.  Just expand all the PIC macros for clarity.