Add symbolic traceback implementation
The procfs method of searching for elf symbols in the program map is not
portable to FreeBSD.  Instead, use popen to the addr2line program to
translate addresses one by one.  As logging is a diagnostic tool that is
off by default, this slow implementation is acceptable.

As a bonus, we get a lot more information about the stack point than
is given by solaris Implementation.